+++ packages/cli/src/commands/generate.ts
@@ -302,9 +302,7 @@ async function outputCodegen(
   )
 
   const hash = h64(
-    Binary.fromText(
-      Array.from(metadataTypes.checksumToIdx.keys()).join(""),
-    ).asBytes(),
+    Binary.fromText(Array.from(metadataTypes.checksumToIdx.keys()).join("")),
   )
 
   const EntryPointsCodec = Vector(EntryPointCodec)