+++ packages/client/src/tx/types.ts
@@ -1,12 +1,7 @@
 import { PullOptions } from "@/types"
 import { SystemEvent } from "@polkadot-api/observable-client"
 import { PolkadotSigner } from "@polkadot-api/polkadot-signer"
-import {
-  Binary,
-  Enum,
-  HexString,
-  SS58String,
-} from "@polkadot-api/substrate-bindings"
+import { Enum, HexString, SS58String } from "@polkadot-api/substrate-bindings"
 import { Observable } from "rxjs"
 
 export type TxEvent = TxSigned | TxBroadcasted | TxBestBlocksState | TxFinalized
@@ -247,7 +242,7 @@ export interface TxCall {
    *
    * @returns Promise resolving in the encoded data.
    */
-  (): Promise<Binary>
+  (): Promise<Uint8Array>
 }
 
 export interface TxBare {
@@ -413,7 +408,7 @@ export type OfflineTxEntry<
   /**
    * SCALE-encoded callData of the transaction.
    */
-  encodedData: Binary
+  encodedData: Uint8Array
   /**
    * PAPI way of expressing an extrinsic with arguments.
    * It's useful to pass as a parameter to extrinsics that accept calls.
@@ -430,7 +425,7 @@ export type TxFromBinary<Asset> = {
    * @returns Transaction object.
    */
   (
-    callData: Binary,
+    callData: Uint8Array,
     options?: PullOptions,
   ): Promise<Transaction<any, string, string, Asset>>
 }