+++ packages/client/src/typed-codecs/shaped-codec.ts
@@ -1,4 +1,4 @@
-import { Binary, Codec, ResultPayload } from "@polkadot-api/substrate-bindings"
+import { Codec, ResultPayload } from "@polkadot-api/substrate-bindings"
 
 type Tuple<T> = [T, ...T[]]
 type UnionToIntersection<U> = (U extends any ? (k: U) => void : never) extends (
@@ -30,7 +30,6 @@ type _Inner<T> = T extends void
           | null
           | symbol
           | Uint8Array
-          | Binary
       ? {}
       : T extends Tuple<any>
         ? {