+++ packages/codegen/CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Changed
 
 - Updated generated docs to match v2
+- Codegen exposes and expects `Vec<u8>` as `Uint8Array` instead of `Binary`.
 
 ## Unreleased